Added LaTeX compile trigger from URL with &compile:pdftex and &compile:xetex triggers
