+++ src/components/editor/Editor.tsx
@@ -17,9 +17,9 @@ import { pluginRegistry } from "../../plugins/PluginRegistry";
 import { EditorLoader } from "../../services/EditorLoader.ts";
 import { fileStorageService } from "../../services/FileStorageService";
 import type { DocumentList } from "../../types/documents.ts";
-import { buildUrlWithFragments, parseUrlFragments } from "../../types/yjs";
+import { buildUrlWithFragments, parseUrlFragments } from "../../utils/urlUtils";
 import { copyCleanTextToClipboard } from "../../utils/clipboardUtils";
-import { fileCommentProcessor } from "../../utils/fileCommentProcessor.ts";
+import { fileCommentProcessor } from "../../utils/fileCommentProcessor";
 import { arrayBufferToString } from "../../utils/fileUtils";
 import CommentPanel from "../comments/CommentPanel";
 import CommentToggleButton from "../comments/CommentToggleButton";
